+++ src/Network/Network/ChangeLog.md
@@ -58,6 +58,10 @@
 * Deleted NetworkWatcher cmdlets for "Classic" Connection Monitor
     - `Start-AzNetworkWatcherConnectionMonitor`
     - `Get-AzNetworkWatcherConnectionMonitorReport`
+* Updated cmdlets to support topology-specific connectivity capabilities:
+    - `New-AzNetworkManagerConnectivityConfiguration`
+    - `Set-AzNetworkManagerConnectivityConfiguration`
+    - `Get-AzNetworkManagerConnectivityConfiguration`
 
 ## Version 7.15.1
 * Updated VirtualNetworkGatewayConnection cmdlets to pass AuxilaryAuthHeader for referenced resourceIds i.e. LocalNetworkGateway2, VirtualNetworkGateway2. This is needed in case referenced resourceIds are in different AAD Tenant.